Show buttons in session-less state as well
